Add files via upload
Project summary: 

Examine the various security features, especially Non-Repudiation, and designed a proof of concept for maintaining the true OFE in medical data exchange using modern cryptographic techniques such as Digital Signature and Asymmetric Encryption with Python. Developed a Blockchain web application with Nodejs and React to facilitate the interchange of medical records thus addressing the limitation of both 3 non-Blockchain-based and other Blockchain-based techniques. Performed validation and testing of the model to support the adequacy of the concept.